backend/config.py
import os
from dotenv import load_dotenv

load_dotenv()

# Configuração do banco de dados
DATABASE_URL = os.getenv(
    "DATABASE_URL",
    "postgresql+psycopg2://user:password@localhost:5432/chatbot_rag"
)

# Configuração de embeddings (Sentence Transformers)
# all-MiniLM-L6-v2: rápido e eficiente, 384 dimensões
EMBEDDING_MODEL = os.getenv("EMBEDDING_MODEL", "all-MiniLM-L6-v2")
EMBEDDING_DIMENSION = 384
